test(node/manager): consolidate concurrency/benchmark tests into manager_test.go
Merge manager_lock_test.go and manager_concurrency_test.go into the single
manager_test.go. All of the lock/benchmark content (benchK8s/benchWorker +
BenchmarkAddNode_Concurrent, countingWorker + newConcurrencyManager, gateK8s +
TestAddNode_LockFreePathRunsInParallel, and the same-node/distinct-nodes/
add-update-delete concurrency tests) moves over unchanged.

From manager_concurrency_test.go, drop Test_ConcurrentAddSameNode_SingleSubmit
(overlaps TestAddNode_ConcurrentSameNode_SingleJob) and Test_ConcurrentNodeOps
(its Add/Update/Delete hammer overlaps TestNode_ConcurrentAddUpdateDelete_NoRace;
its only distinct dimension, concurrent GetNode, is covered by keeping
Test_ConcurrentGetNode_NoRace). The go-deadlock detector still exercises the
lock under load via the surviving concurrency tests.
